PermissionValidator will now check the rank change of a group.
It will check if the rankFrom is equal to the given name of the rank in
the config. If this would not be equal that would not work. 

It also checks if rankTo exists.

If one or both of the checks are false, a message will be displayed and
Autorank will disable itself to protect admins.