Add new UUID storage system:
This will store all uuids ever looked up so we don't have to rely on the
'internet lookup' that heavily anymore. All the uuids will be stored
once a player joins, and will be valid for 24 hours. Every time a player
joins, it will check if the stored uuid is out of date. If it is out of
date, it will be refreshed with a new one. Meanwhile, most of Autorank
will use the UUID storage system, so that there are no more blocking
calls.

Signed-off-by: Staartvin <mijnleraar@msn.com>