Few changes:
Organize imports
Fix formats
Switch over to getStoredUUID() instead of getUUIDOfPlayer(). This is
done as the latter one can query the Mojang website, which we do not
want to happen that often. Therefore Autorank mostly checks its cached
UUIDs. Once a player logs in, Autorank will update the UUID record of
that player (if needed).